Restructure quantizers
